moved type-related lines to types.go
